Add prototype code entropy loss

+++ scripts/train_vq_code_predictor.py
@@ -6,17 +6,22 @@ parser = argparse.ArgumentParser(
     description="MLP code prediction trainer",
 )
 parser.add_argument(
- "--data_dir", type=str, metavar="PATH", default='./data/', help="Path to data folder"
+ "--data_dir", type=str,  default='./data/', help="Path to data folder"
 )
 parser.add_argument(
- "--model_path", type=str, metavar="PATH", default='./runs/sep_ae/20201223_183223_vae_wa_6h_quantized_256_16qh/', help="Path to data folder"
+ "--model_path", type=str,  default='./runs/sep_ae/20201230_132811_vae_wa_6h_quantized_256_16qh_chunk-drop30/', help="Path to model folder"
 )
 parser.add_argument(
- "--output_path", type=str, metavar="PATH", default='./runs/mlpcodepredictor/', help="Path to data folder"
+ "--output_path", type=str,  default='./runs/mlpcodepredictor/', help="Path to output folder"
+)
+
+parser.add_argument(
+ "--dataset", type=str,  default='wikianswers', help="Which dataset?"
 )
 
 parser.add_argument("--train", action="store_true", help="Train mode")
 parser.add_argument("--eval", action="store_true", help="Eval mode")
+parser.add_argument("--test", action="store_true", help="Eval on test")
 
 parser.add_argument(
  "--lr", type=float, default=1e-4
@@ -34,8 +39,21 @@ parser.add_argument(
  "--num_steps", type=int, default=30001
 )
 
+
+
 args = parser.parse_args()
 
+if args.dataset == 'wikianswers':
+    dataset_all = 'wikianswers-para-allqs'
+    dataset_clusters = 'wikianswers-pp'
+    dataset_geneval = 'wikianswers-para-splitforgeneval'
+    dataset_mlppredict = 'wikianswers-para-exemplarmlppredict'
+elif args.dataset == 'qqp':
+    dataset_all = 'qqp-allqs'
+    dataset_clusters = 'qqp-clusters'
+    dataset_geneval = 'qqp-splitforgeneval'
+    dataset_mlppredict = 'qqp-exemplarmlppredict'
+    
 
 import torch
 from torch.autograd import Variable
@@ -76,7 +94,7 @@ import jsonlines, os
 
 MODEL_PATH = args.model_path
 
-if not os.path.exists(MODEL_PATH+f'/sep_encoding_1_train.npy'):
+if not os.path.exists(MODEL_PATH+f'/sep_encoding_1_train.npy') or not os.path.exists(MODEL_PATH+f'/sep_encoding_1_dev.npy') or not os.path.exists(MODEL_PATH+f'/sep_encoding_1_test.npy'):
     # generate encodings
     print('Encoding cache not found - generating...')
 
@@ -99,7 +117,7 @@ if not os.path.exists(MODEL_PATH+f'/sep_encoding_1_train.npy'):
         cfg_dict["training"]["dataset"] = 'json'
         cfg_dict["training"]["shuffle_data"] = False
         cfg_dict['json_dataset'] = {
-            "path": "wikianswers-para-allqs",
+            "path": dataset_all,
             "field_map": [
                 {
                     "type": "copy",
@@ -133,28 +151,37 @@ if not os.path.exists(MODEL_PATH+f'/sep_encoding_1_train.npy'):
     instance.load_checkpoint(checkpoint_path)
     instance.model.eval()
 
-    # _, _, _, memory_train = instance.validate(data_loader, use_train=True, save_model=False, memory_keys_to_return=['sep_encoding_1', 'sep_encoding_2','vq_codes'])
-    _, _, _, memory_train = instance.inference(data_loader.train_loader, memory_keys_to_return=['sep_encoding_1', 'sep_encoding_2','vq_codes'])
+    # Train
+    if not os.path.exists(MODEL_PATH+f'/sep_encoding_1_train.npy'):
+        _, _, _, memory_train = instance.inference(data_loader.train_loader, memory_keys_to_return=['sep_encoding_1', 'sep_encoding_2','vq_codes'])
 
-    # print(len(memory_train))
+        torch.cuda.empty_cache()
 
-    torch.cuda.empty_cache()
+        for mem_key in ['sep_encoding_1', 'sep_encoding_2','vq_codes']:
+            np.save(MODEL_PATH+f'/{mem_key}_train.npy', memory_train[mem_key])
+    
+    # Dev
+    if not os.path.exists(MODEL_PATH+f'/sep_encoding_1_dev.npy'):
+        _, _, _, memory_dev = instance.inference(data_loader.valid_loader, memory_keys_to_return=['sep_encoding_1', 'sep_encoding_2','vq_codes'])
 
-    for mem_key in ['sep_encoding_1', 'sep_encoding_2','vq_codes']:
-        np.save(MODEL_PATH+f'/{mem_key}_train.npy', memory_train[mem_key])
-            
-    # _, _, _, memory_dev = instance.validate(data_loader, use_train=False, save_model=False, memory_keys_to_return=['sep_encoding_1', 'sep_encoding_2','vq_codes'])
-    _, _, _, memory_dev = instance.inference(data_loader.valid_loader, memory_keys_to_return=['sep_encoding_1', 'sep_encoding_2','vq_codes'])
+        torch.cuda.empty_cache()
 
-    # print(len(memory_dev))
+        for mem_key in ['sep_encoding_1', 'sep_encoding_2','vq_codes']:
+            np.save(MODEL_PATH+f'/{mem_key}_dev.npy', memory_dev[mem_key])
+    
+    # Test
+    if not os.path.exists(MODEL_PATH+f'/sep_encoding_1_test.npy'):
+        _, _, _, memory_test = instance.inference(data_loader.test_loader, memory_keys_to_return=['sep_encoding_1', 'sep_encoding_2','vq_codes'])
 
-    torch.cuda.empty_cache()
+        torch.cuda.empty_cache()
 
-    for mem_key in ['sep_encoding_1', 'sep_encoding_2','vq_codes']:
-        np.save(MODEL_PATH+f'/{mem_key}_dev.npy', memory_dev[mem_key])
+        for mem_key in ['sep_encoding_1', 'sep_encoding_2','vq_codes']:
+            np.save(MODEL_PATH+f'/{mem_key}_test.npy', memory_test[mem_key])
 
     del instance
     del data_loader
+
+    torch.cuda.empty_cache()
 
     print('Encoding cache built')
 
@@ -164,12 +191,17 @@ print('Loading encodings, data')
 
 memory_train = {}
 memory_dev = {}
+memory_test = {}
 
 for mem_key in ['sep_encoding_1', 'sep_encoding_2', 'vq_codes']:
     memory_train[mem_key] = np.load(MODEL_PATH+f'/{mem_key}_train.npy')
-    memory_dev[mem_key] = np.load(MODEL_PATH+f'/{mem_key}_dev.npy')
     
-with jsonlines.open(os.path.join(args.data_dir, "wikianswers-pp/train.jsonl")) as f:
+    if args.test:
+        memory_test[mem_key] = np.load(MODEL_PATH+f'/{mem_key}_test.npy')
+    else:
+        memory_dev[mem_key] = np.load(MODEL_PATH+f'/{mem_key}_dev.npy')
+    
+with jsonlines.open(os.path.join(args.data_dir, dataset_clusters, "train.jsonl")) as f:
     train_qs = [row for row in f]
 train_cluster_ixs = []
 ix = 0
@@ -177,11 +209,12 @@ for cix, cluster in enumerate(train_qs):
     clen = len(cluster['qs'])
     for i in range(clen):
         cluster_ixs = list(range(ix, ix+clen))
+        # if args.dataset != 'qqp':
         cluster_ixs.remove(ix + i)
         train_cluster_ixs.append(cluster_ixs)
     ix += clen
     
-with jsonlines.open(os.path.join(args.data_dir, "wikianswers-pp/dev.jsonl")) as f:
+with jsonlines.open(os.path.join(args.data_dir, dataset_clusters, "dev.jsonl")) as f:
     dev_qs = [row for row in f]
 dev_cluster_ixs = []
 ix = 0
@@ -189,11 +222,26 @@ for cix, cluster in enumerate(dev_qs):
     clen = len(cluster['qs'])
     for i in range(clen):
         cluster_ixs = list(range(ix, ix+clen))
+        # if args.dataset != 'qqp':
         cluster_ixs.remove(ix + i)
         dev_cluster_ixs.append(cluster_ixs)
     ix += clen
 
+import sys, gc
+gc.collect()
+# print('mem train', sum([x.nbytes for x in memory_train.values()])/1024**2)
+# print('mem dev', sum([x.nbytes for x in memory_dev.values()])/1024**2)
+# print('mem test', sum([x.nbytes for x in memory_test.values()])/1024**2)
+# print('qs train', sys.getsizeof(train_qs)/1024**2)
+# print('qs dev', sys.getsizeof(dev_qs)/1024**2)
+# print('clusters train', sys.getsizeof(train_cluster_ixs)/1024**2)
+# print('clusters dev', sys.getsizeof(dev_cluster_ixs)/1024**2)
+
 print('Data and encodings loaded')
+
+# from guppy import hpy; 
+# h=hpy()
+# h.heap()
     
 
 # Prepare datasets
@@ -204,25 +252,36 @@ h_ix = 0
 X = np.concatenate([memory_train['sep_encoding_1'][:, 0, :], memory_train['sep_encoding_2'][:, 0, :]], axis=1)
 y = memory_train['vq_codes'][:, :, 0]
 
-X_train_ixs = []
-y_train_ixs = []
-for src_ix, cluster in enumerate(train_cluster_ixs):
-    for tgt_ix in cluster:
-        X_train_ixs.append(src_ix)
-        y_train_ixs.append(tgt_ix)
+# print(y[:10, :])
+# print(len(train_qs))
+# print(X.shape)
+# print(len(train_cluster_ixs))
+
+# X_train_ixs = []
+# y_train_ixs = []
+# for src_ix, cluster in enumerate(train_cluster_ixs):
+#     for tgt_ix in cluster:
+#         X_train_ixs.append(src_ix)
+#         y_train_ixs.append(tgt_ix)
         
 
-X_dev_ixs = []
-y_dev_ixs = []
-for src_ix, cluster in enumerate(dev_cluster_ixs[:1000]):
-    for tgt_ix in cluster:
-        X_dev_ixs.append(src_ix)
-        y_dev_ixs.append(tgt_ix)
+# X_dev_ixs = []
+# y_dev_ixs = []
+# for src_ix, cluster in enumerate(dev_cluster_ixs[:1000]):
+#     for tgt_ix in cluster:
+#         X_dev_ixs.append(src_ix)
+#         y_dev_ixs.append(tgt_ix)
 
-# X_dev = memory_dev['sep_encoding_1'][:, 0, :]
-X_dev = np.concatenate([memory_dev['sep_encoding_1'][:, 0, :], memory_dev['sep_encoding_2'][:, 0, :]], axis=1)
-y_dev = memory_dev['vq_codes'][:, :, 0]
 
+if args.test:
+    # X_dev = memory_dev['sep_encoding_1'][:, 0, :]
+    X_test = np.concatenate([memory_test['sep_encoding_1'][:, 0, :], memory_test['sep_encoding_2'][:, 0, :]], axis=1)
+    y_test = memory_test['vq_codes'][:, :, 0]
+else:
+
+    # X_dev = memory_dev['sep_encoding_1'][:, 0, :]
+    X_dev = np.concatenate([memory_dev['sep_encoding_1'][:, 0, :], memory_dev['sep_encoding_2'][:, 0, :]], axis=1)
+    y_dev = memory_dev['vq_codes'][:, :, 0]
 print('Datasets prepped')
 
 # Train the model
@@ -263,6 +322,8 @@ if args.train:
         
         inputs = Variable(torch.tensor([X[ix] for ix in batch_ixs])).cuda()
         
+        # print([len(train_cluster_ixs[cix]) for cix in batch_ixs])
+
         tgt = torch.where(torch.cat([torch.sum(torch.cat([onehot(torch.tensor(y[ix]), N=output_dim).unsqueeze(0) for ix in train_cluster_ixs[cix]], dim=0), dim=0, keepdims=True) for cix in batch_ixs], dim=0) > 0, 1, 0).cuda()
         
     #     tgt = Variable(tgt).cuda()
@@ -303,9 +364,10 @@ if args.train:
             accuracy = 100 * correct/(total*NUM_HEADS)
             head_acc = [100*x/total for x in head_acc]
             all_accuracy = 100 * all_acc/total
-            if all_accuracy > best_acc:
+            if accuracy > best_acc:
+                print('Saving...')
                 torch.save(model.state_dict(), args.output_path+'/code_predict.pt')
-                best_acc = all_accuracy
+                best_acc = accuracy
                 metrics = {
                     'acc': accuracy, 
                     'full_acc': all_accuracy, 
@@ -321,7 +383,9 @@ if args.train:
 
 
 # Run inference
-if args.eval:
+if args.eval or args.test:
+
+    split = 'test' if args.test else 'dev'
 
     print('Generating exemplars')
 
@@ -330,19 +394,18 @@ if args.eval:
 
     NUM_HEADS = 16
     NUM_TEMPL_HEADS = 4
-    CODEBOOK_SIZE = 64
 
     model.load_state_dict(torch.load(args.output_path+'/code_predict.pt'))
     model.eval()
 
 
 
-    with jsonlines.open(os.path.join(args.data_dir, "wikianswers-para-splitforgeneval/dev.jsonl")) as f:
+    with jsonlines.open(os.path.join(args.data_dir, f"{dataset_geneval}/{split}.jsonl")) as f:
         rows = [row for row in f]
         
     q_to_ix = {}
     ix = 0
-    with jsonlines.open(os.path.join(args.data_dir, "wikianswers-pp/dev.jsonl")) as f:
+    with jsonlines.open(os.path.join(args.data_dir, f"{dataset_clusters}/{split}.jsonl")) as f:
         dev_qs = [row for row in f]
     for cix, cluster in enumerate(dev_qs):
         for q in cluster['qs']:
@@ -375,44 +438,43 @@ if args.eval:
     #         this_row['vq_codes'] = tgt_codes
     #         f.write(this_row)
 
-    os.makedirs(args.data_dir + '/wikianswers-para-exemplarmlppredict', exist_ok=True)
-    with jsonlines.open(args.data_dir + '/wikianswers-para-exemplarmlppredict/dev.jsonl', 'w') as f:
+    X_src = X_test if args.test else X_dev
+
+    os.makedirs(args.data_dir + '/' + dataset_mlppredict, exist_ok=True)
+    with jsonlines.open(args.data_dir + '/' + dataset_mlppredict +f'/{split}.jsonl', 'w') as f:
 
         for ix, row in enumerate(tqdm(rows)):
             query_ix = q_to_ix[row['sem_input']]
             tgt_codes = [0] * (NUM_HEADS - NUM_TEMPL_HEADS)
 
-            inputs = Variable(torch.tensor([X_dev[query_ix]])).cuda()
+            inputs = Variable(torch.tensor([X_src[query_ix]])).cuda()
 
             outputs = model(inputs)[0]
-            probs, predicted = torch.topk(torch.softmax(outputs, -1), 6 -1)
+            probs, predicted = torch.topk(torch.softmax(outputs, -1), 3 -1)
 
     #         print(predicted.shape, probs.shape)
         #     break
 
             joint_probs = [([], 0)]
-
             for h_ix in range(NUM_TEMPL_HEADS):
                 new_hypotheses = []
                 for i, (combo, prob) in enumerate(joint_probs):
-                    for k in range(5):
+                    for k in range(2):
                         new_hyp = [copy.copy(combo), prob]
                         new_hyp[0].append(predicted[h_ix, k].item())
                         new_hyp[1] += torch.log(probs[h_ix, k]).item()
-        #                 print(combo, prob)
 
                         new_hypotheses.append(new_hyp)
-        #                 break
 
                 joint_probs = new_hypotheses
-                joint_probs = sorted(joint_probs, key=lambda x: x[1], reverse=True)[:10]
-        #     break
-
-            gold = y_dev[ix]
-
-        #         print(predicted, gold)
-            pred_codes = [tgt_codes + x[0] for x in sorted(joint_probs, key=lambda x: x[1], reverse=True)[:5]]
-    #         print(pred_codes)
+                joint_probs = sorted(joint_probs, key=lambda x: x[1], reverse=True)[:3]
+            pred_codes = [tgt_codes + x[0] for x in sorted(joint_probs, key=lambda x: x[1], reverse=True)[:2]]
+            
+            # pred_codes = predicted.transpose(1,0).tolist()
+            # pred_codes = [tgt_codes + codes for codes in pred_codes]
+            # print(pred_codes)
+            # exit()
+    
 
         #     for h_ix in range(NUM_TEMPL_HEADS):
         #         tgt_codes.append(predicted[0, h_ix].item())
